moved back util methods to avoid changes in rules
