+++ diktat-api/src/main/kotlin/com/saveourtool/diktat/common/config/rules/LegacyAliases.kt
