+++ diktat-rules/src/main/kotlin/com/saveourtool/diktat/ruleset/utils/AstNodeUtils.kt
@@ -12,7 +12,7 @@
 
 package com.saveourtool.diktat.ruleset.utils
 
-import com.saveourtool.diktat.common.config.rules.DIKTAT
+import com.saveourtool.diktat.DIKTAT
 import com.saveourtool.diktat.common.config.rules.Rule
 import com.saveourtool.diktat.common.config.rules.RulesConfig
 import com.saveourtool.diktat.common.config.rules.isAnnotatedWithIgnoredAnnotation
